Server endpoints completed

+++ Winter/src/main/java/com/revature/repository/CategoryRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.revature.model.Category;
 @Repository
-public interface CategoryRepo extends JpaRepository<Category, Integer>{
+public interface CategoryRepo extends JpaRepository<Category, Long>{
 
 }
